requirements.txt
streamlit
numpy
deta
Keras-Preprocessing
keras-models
tensorflow
